+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## Release 1.1.0
 
+**Features**
+- configure logging directory for instances
+
 **Fixed**
 - no shared etc & instances dir over different Karaf versions, after upgrade instances need to be build from scratch