+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 **Fixed**
 - no shared etc & instances dir over different Karaf versions, after upgrade instances need to be build from scratch
 - instance defaults via params
+- fix SELinux types of configuration files
 
 ## Release 1.0.1