+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## Release 1.1.1
+
+**Fixed**
+- update reference.md to contain latest additions
+
 ## Release 1.1.0
 
 **Features**